AoE Bonus Additions
Various abilities that weren't compatible with AoE bonuses now function with them.
This affects any ability with a 'radius' value that didn't already have it, but also ones with 'width' as well.

This change is easily reversible, but I hope I didn't overstep my bounds. I figured this was easy enough to do.